fix: add transformError to deleteOne and custom methods

+++ packages/rest/src/create-data-provider.ts
@@ -268,9 +268,15 @@ export const createDataProvider = (
           searchParams: qs.stringify(query, { encodeValuesOnly: true }),
         });
 
-        const data = await options.deleteOne.mapResponse(response, params);
+        if (response.ok) {
+          const data = await options.deleteOne.mapResponse(response, params);
 
-        return { data };
+          return { data };
+        }
+
+        const error = await options.deleteOne.transformError(response, params);
+
+        throw error;
       },
       deleteMany: options.deleteMany
         ? async (params): Promise<DeleteOneResponse<any>> => {
@@ -331,9 +337,15 @@ export const createDataProvider = (
 
         const response = await client<AnyObject>(url);
 
-        const data = await options.custom.mapResponse(response, params);
+        if (response.ok) {
+          const data = await options.custom.mapResponse(response, params);
 
-        return { data };
+          return { data };
+        }
+
+        const error = await options.custom.transformError(response, params);
+
+        throw error;
       },
       getApiUrl: () => apiURL,
     },
